rename factory -> benchmark_factory
